Player custumisation (#18)
* basics

Added the basics of player customisation. Will also be tweaking the
movement in this branch to make it "feel" better.

* some improvements

Mostly bugs, tweaking and code cleaning.

*  demo for customisation, improvements to movement

added some temporary sprites for the customisation and fixed some issues
relating to it. Also added improvements to the movement system.

* final commit

just changed the room views setting to center it on the player